refactor(gemini): split provider tests for GH727
